pyproject.toml
[project]
name = "werewolf-state-machine"
version = "0.1.0"
description = "Gameplay for the social deception game Werewolf"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
dependencies = [
    "loguru>=0.7.3",
    "nats-py>=2.11.0",
    "pydantic>=2.11.7",
    "python-statemachine>=2.5.0",
    "typer>=0.16.0",
]

[project.scripts]
werewolf = "werewolf:app"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
